Overall: All tables now created correctly.
time changed to arrivalTime because time is an optionally reserved word.
level changed to severity because level is a reserved word.
write? changed to writePermission because ? is an invalid character.
date attribute in RecordedTransaction changed to transactionDate and date attribute in TracksDate changed to violationDate because date is a reserved word.

DATETIME datatype changed to TIMESTAMP because Oracle does not have a DATETIME datatype. BOOLEAN datatype changed to NUMBER(1) because Oracle does not support BOOLEAN datatype. Removed colons from attribute declarations.

ON UPDATE CASCADE commented because Oracle does not support it.

Deleted UNIQUE constraint on startTime in ScheduledShift because would cause a reference loop with ScheduledTime. Seems to be vestigial.

Deleted Consumables table. Helps resolve issue with only cid from composite primary key being referred to by Perishables and nonPerishables.

Wrapped primary and foreign key declarations in parentheses and added commas as necessary.